# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "<h2>To be able to bypass the JWT authentication for the endpoints needing authorization,         login first, then use the token recieved, press the <b>Authorize</b> button and type 'Bearer ' and paste the token copied.         After, you can access the enpoints requiring authorization without filling the <b>Authorization</b> field.</h2>"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
...
    "/orders/": {
      "get": {
        "tags": [
          "The Order & Order Item Tables Endpoints"
        ],
        "description": "<h3>Gets aunthenticated user's all orders with all the order items or if the authenticated user has an admin role then all the users' order history with the related order items.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Order & Order Item Tables Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, if given ordernumber exists, makes changes to order's status based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/orderStatusUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orderstatuses/": {
      "get": {
        "tags": [
          "The Order Status Table Endpoints"
        ],
        "description": "<h3>Gets the order status list.</h3>",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "The Order Status Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, creates an order status based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/orderStatusCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Order Status Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, changes an order status based on validated parameters provided in the request's body. This only happens if it is not a default order status.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/orderStatusChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "The Order Status Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, removes an order status based on validated parameters provided in the request's body. This happens only if it is not a default order status or there are no orders with this order status.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/orderStatusDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/orders": {
      "get": {
        "tags": [
          "The Admin Front-End Endpoints"
        ],
        "description": "<h3>Renders the orders page if the token owner has an admin role. If not admin then returns the login page back to the user.</h3>",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "orderStatusUpdate": {
      "type": "object",
      "properties": {
        "ordernumber": {
          "type": "string",
          "example": "qvj79w0b"
        },
        "orderstatusid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "ordernumber",
        "orderstatusid"
      ]
    },
    "orderStatusCreate": {
      "type": "object",
      "properties": {
        "orderstatus": {
          "type": "string",
          "example": "Ordered"
        }
      },
      "required": [
        "orderstatus"
      ]
    },
    "orderStatusChange": {
      "type": "object",
      "properties": {
        "orderstatusid": {
          "type": "number",
          "example": 2
        },
        "orderstatus": {
          "type": "string",
          "example": "Completed"
        }
      },
      "required": [
        "orderstatusid",
        "orderstatus"
      ]
    },
    "orderStatusDelete": {
      "type": "object",
      "properties": {
        "orderstatusid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "orderstatusid"
      ]
    }
  }
}